redirect page crud user on role not super-admin

+++ app/Http/Controllers/UserController.php
@@ -13,16 +13,26 @@ class UserController extends Controller
 {
     public function index()
     {
-        return view('dashboard.user', [
-            'judul_halaman' => 'Admin | Data Pengguna',
-            'users' => User::latest()->get()
-        ]);
+        if( Auth::user()->role === 'super-admin' ) {
+            return view('dashboard.user', [
+                'judul_halaman' => 'Admin | Data Pengguna',
+                'users' => User::latest()->get()
+            ]);
+        }
+        else{
+            return redirect('/dashboard');
+        }
     }
     public function create()
     {
-        return view('dashboard.add-user', [
-            'judul_halaman' => 'Admin | Tambah Pengguna'
-        ]);
+        if( Auth::user()->role === 'super-admin' ) {
+            return view('dashboard.add-user', [
+                'judul_halaman' => 'Admin | Tambah Pengguna'
+            ]);
+        }
+        else{
+            return redirect('/dashboard');
+        }
     }
     public function store(Request $request)
     {
@@ -64,18 +74,28 @@ class UserController extends Controller
 
     public function show(User $user)
     {
-        return view('dashboard.detail-user', [
-            'judul_halaman' => 'Admin | Detail Pengguna',
-            'user' => $user
-        ]);
+        if( Auth::user()->role === 'super-admin' ) {
+            return view('dashboard.detail-user', [
+                'judul_halaman' => 'Admin | Detail Pengguna',
+                'user' => $user
+            ]);
+        }
+        else{
+            return redirect('/dashboard');
+        }
     }
 
     public function edit(User $user)
     {
-        return view('dashboard.edit-user', [
-            'judul_halaman' => 'Admin | Edit Pengguna',
-            'user' => $user
-        ]);
+        if( Auth::user()->role === 'super-admin' ) {
+            return view('dashboard.edit-user', [
+                'judul_halaman' => 'Admin | Edit Pengguna',
+                'user' => $user
+            ]);
+        }
+        else{
+            return redirect('/dashboard');
+        }
     }
     public function update(Request $request, User $user)
     {
@@ -161,10 +181,15 @@ class UserController extends Controller
 
     public function destroy(User $user)
     {
-        if ($user->foto_profil) {
-            Storage::delete($user->foto_profil);
+        if( Auth::user()->role === 'super-admin' ) {
+            if ($user->foto_profil) {
+                Storage::delete($user->foto_profil);
+            }
+            User::destroy($user->id);
+            return redirect('/dashboard/user')->with('success', 'menghapus');
         }
-        User::destroy($user->id);
-        return redirect('/dashboard/user')->with('success', 'menghapus');
+        else {
+            return redirect('/dashboard');
+        }
     }
 }
